requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Pydantic for data validation
pydantic==2.5.0
pydantic[email]==2.5.0

# Google Gemini AI integration via LangChain
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-core==0.1.0
google-generativeai==0.3.2

# Document processing libraries
PyMuPDF==1.23.8
pdfplumber==0.10.3
docx2txt==0.8
python-docx==1.1.0

# OCR support
pytesseract==0.3.10
Pillow==10.1.0

# Machine learning and text analysis
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.4

# NLP processing
spacy==3.7.2
nltk==3.8.1

# HTTP client for external API calls
httpx==0.25.2
requests==2.31.0

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
click==8.1.7
tqdm==4.66.1
python-magic==0.4.27
chardet==5.2.0

# Database support (if needed for future features)
sqlalchemy==2.0.23
alembic==1.13.1

# Caching (if needed)
redis==5.0.1

# File handling
aiofiles==23.2.1

# Date and time handling
python-dateutil==2.8.2

# Environment and configuration
pyyaml==6.0.1
toml==0.10.2

# Security
cryptography==41.0.8
passlib==1.7.4

# Async support
asyncio==3.4.3
aiohttp==3.9.1

# Text processing
textstat==0.7.3
wordcloud==1.9.2

# PDF manipulation
reportlab==4.0.7
